Update ProvTable file structure
