+++ context/app/static/js/components/Detail/ProvTable/index.js
@@ -1,0 +1,3 @@
+import ProvTable from './ProvTable';
+
+export default ProvTable;
